refactor(grader): default_tune_description lives next to TaskGrader; hoist imports
The default tune description belongs with the grader contract it defaults,
not in the templating module. Moving it pulls SubprocessGrader's tiny
indirection wrapper out of the picture too — it was importing from
coral.template.coral_md, which was importing from itself.

- New `DEFAULT_TUNE_DESCRIPTION` constant + `default_tune_description()`
  function at the top of `coral/grader/task_grader.py` (single source
  of truth).
- `TaskGrader.describe_tune()` now returns the constant.
- `coral.template.coral_md` re-exports `default_tune_description` for
  back-compat with any caller that already imported from there, but the
  duplicate inline definition is gone.
- `SubprocessGrader` drops its `_default_tune_description` wrapper and
  imports the canonical name directly.
- `coral.agent.manager` also imports from the canonical location.

Hoist function-local imports to module top in the same files (per
project style):
- `task_grader.py`: `shutil`, `sys` → top
- `subprocess_grader.py`: `Score` → top (was duplicated inside the
  TimeoutExpired branch)
- `agent/manager.py`: `load_grader` → top; remove redundant `import
  time` shadowing the module-level one

Pre-existing lazy imports with rationale comments (CLI subcommand
dispatch in `cli/__init__.py`, `cli/start.py`, `cli/eval.py`; the
"don't pull in grader on CLI-only paths" pattern in `manager.py`)
are left as-is — those have a deliberate design reason that isn't
a coding-style issue.

199/199 tests pass; ruff check + ruff format both clean on PR files.
